switch possition of liseners to drag file
